refactor(viewer): extract viewer_notifications.jsx from viewer.jsx
Pure file split, no behavior change. Step 8 of mp-pxxc.
