+++ web-mobile/index.html
@@ -83,6 +83,10 @@
          MyMatchAlertBanner, isFollowedMatchOnDeck), extracted from viewer.jsx (mp-pxxc step 3).
          Must load after viewer_watchlist_core.js and viewer_watchlist.js, and before viewer.js. -->
     <script type="module" src="/dist/viewer_alerts.js?v=5"></script>
+    <!-- Notification-settings + announcement UI (notificationSupported, AnnBellBtn,
+         AnnouncementCard, AnnouncementBanner), extracted from viewer.jsx (mp-pxxc step 8).
+         Must load after viewer_alerts.js (its dep) and before viewer.js. -->
+    <script type="module" src="/dist/viewer_notifications.js?v=5"></script>
     <!-- Match-display components (VSchedItem, MatchDetailCard, MatchViewerModal,
          mymatchQueueLabel, subBoutLabel), extracted from viewer.jsx (mp-pxxc step 4).
          Must load after viewer_utils.js (its dep) and before viewer.js. -->
